refactor(compiler-plugin): refactor the backend
