+++ collektive-compiler-plugin/src/main/kotlin/it/unibo/collektive/compiler/backend/irextensions/DeclarationParent.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2025, Danilo Pianini, Nicolas Farabegoli, Elisa Tronetti,
+ * and all authors listed in the `build.gradle.kts` and the generated `pom.xml` file.
+ *
+ * This file is part of Collektive, and is distributed under the terms of the Apache License 2.0,
+ * as described in the LICENSE file in this project's repository's top directory.
+ */
+
+package it.unibo.collektive.compiler.backend.irextensions
+
+import org.jetbrains.kotlin.ir.declarations.IrDeclarationParent
+import org.jetbrains.kotlin.ir.util.kotlinFqName
+
+
+/**
+ * Returns the fully qualified name of this [IrDeclarationParent] as a string.
+ */
+internal val IrDeclarationParent.fqString get() = this.kotlinFqName.asString()
