+++ collektive-compiler-plugin/src/main/kotlin/it/unibo/collektive/compiler/common/MessageCollectorOps.kt
@@ -6,7 +6,7 @@
  * as described in the LICENSE file in this project's repository's top directory.
  */
 
-package it.unibo.collektive.compiler.utils.logging
+package it.unibo.collektive.compiler.common
 
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
